Puntos del 2 al 9 hechos.

+++ src/main/java/com/alkemy/disney/services/implementaciones/PersonajeServiceImpl.java
@@ -2,14 +2,17 @@ package com.alkemy.disney.services.implementaciones;
 
 import com.alkemy.disney.DTOS.PersonajeBasicDTO;
 import com.alkemy.disney.DTOS.PersonajeDTO;
-
+import com.alkemy.disney.DTOS.PersonajeFiltersDTO;
+import com.alkemy.disney.DTOS.PersonajeUpdateDTO;
 import com.alkemy.disney.entities.PersonajeEntity;
+import com.alkemy.disney.exception.ParamNotFound;
 import com.alkemy.disney.mapper.PersonajeMapper;
 import com.alkemy.disney.repository.PersonajeRepository;
+import com.alkemy.disney.repository.specifications.PersonajeSpecifications;
+import com.alkemy.disney.services.Interfaces.PersonajeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
@@ -22,17 +25,18 @@ public class PersonajeServiceImpl implements PersonajeService {
     PersonajeMapper personajeMapper;
     @Autowired
     PersonajeRepository personajeRepository;
-
+    @Autowired
+    PersonajeSpecifications personajeSpecifications;
 
     @Override
     public PersonajeDTO save(PersonajeDTO dto) {
+        //No se puede crear un personaje con el mismo nombre
         Optional<PersonajeEntity> personaje = personajeRepository.findByNombre(dto.getNombre());
+
         if (personaje.isPresent()) {
-            return null;// TODO:Implementar excepcion
+            throw new ParamNotFound("El personaje ya existe");
         }
-        else{
-
-
+        else {
             PersonajeEntity entity = personajeMapper.personajeDTO2Entity(dto);
             PersonajeEntity personajeSave = personajeRepository.save(entity);
             PersonajeDTO res = personajeMapper.personajeEntity2DTO(personajeSave, false);
@@ -41,46 +45,52 @@ public class PersonajeServiceImpl implements PersonajeService {
     }
 
     @Override
-    public Set<PersonajeDTO>  getAll(){
-        Set<PersonajeEntity> entities = new HashSet <PersonajeEntity> () ;
-        List<PersonajeEntity> a =personajeRepository.findAll();
-        for(PersonajeEntity personaje : a)
-            entities.add(personaje);
-            
-        Set<PersonajeDTO> ret= personajeMapper.personajeEntity2DTOList(entities,true);
-        return ret;
+    public PersonajeDTO update(PersonajeUpdateDTO personajeUpdateDTO) {
 
+        if (!personajeRepository.existsById(personajeUpdateDTO.getPersonajeId()))
+            throw new ParamNotFound("El personaje no existe");
+
+        else {
+            PersonajeEntity personaje = personajeRepository.getReferenceById(personajeUpdateDTO.getPersonajeId());
+            PersonajeDTO personajeDTO = personajeMapper.personajeEntity2DTO(personaje, true);
+            PersonajeDTO res = personajeMapper.personajeUpdateDTO2personajeFullDTO(personajeDTO, personajeUpdateDTO);
+            personajeMapper.refreshValues(personaje, res);
+            personajeRepository.save(personaje);
+
+            return res;
+        }
     }
 
     @Override
-    public PersonajeDTO update(PersonajeBasicDTO dto) {
+    public List<PersonajeBasicDTO> getByFilters(String nombre, Integer edad, Set<Long> peliculas) {
+        PersonajeFiltersDTO personajeFiltersDTO = new PersonajeFiltersDTO(nombre, edad, peliculas);
+        List<PersonajeEntity> entities = personajeRepository.findAll(this.personajeSpecifications.getByFilters(personajeFiltersDTO));
+        List<PersonajeBasicDTO> dto = personajeMapper.personajeEntitySet2DTOList(entities);
 
-        Optional<PersonajeEntity> personaje = personajeRepository.findByNombre(dto.getNombre());
-
-        if (!personaje.isPresent()) {
-            return null;// TODO:Implementar excepcion
-        }
-        else{
-
-            PersonajeEntity entity = personajeMapper.personajeBasicDTO2Entity(dto);
-            PersonajeEntity personajeSave = personajeRepository.save(entity);
-            PersonajeDTO res = personajeMapper.personajeEntity2DTO(personajeSave, false);
-            return res;
-        }
-
-
+        return dto;
     }
-
 
     @Override
     public void delete(Long id) {
-       personajeRepository.deleteById(id);
+
+        if (!personajeRepository.existsById(id))
+            throw new ParamNotFound("El personaje no existe");
+        else
+            personajeRepository.deleteById(id);
+
     }
+
     @Override
-    public PersonajeDTO getById(Long id){
-        PersonajeEntity entiti= personajeRepository.getById(id);
-        PersonajeDTO ret= personajeMapper.personajeEntity2DTO(entiti,false);
-        return ret;
+    public PersonajeDTO getById(Long id) {
+
+        if (!personajeRepository.existsById(id))
+            throw new ParamNotFound("El personaje no existe");
+
+        else {
+            PersonajeEntity entiti = personajeRepository.getById(id);
+            PersonajeDTO ret = personajeMapper.personajeEntity2DTO(entiti, true);
+            return ret;
+        }
 
 
     }
